feat: impl. mapping of gstr_2a, gstr_2b
